+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- app: Material components
+- app: Improved demo application
+- ng-bpmn: `autoOpenMinimap` property to automatically open the Minimap (requires `showMinimap` to be `true`)
 
 ## [0.0.2] - 2023-05-09